package/ci: run Linux desktop GL and GLES tests using llvmpipe.
Matches the CI changes done in Magnum. SwiftShader is still used on Mac
GLES3 and on Android, so I can't remove the SKIPs and XFAILs yet, but
thanks to llvmpipe it's now going to be actually tested instead of just
being ran through and then skipped right before drawing anything.