Ui: further increase test thresholds for Mesa 21 llvmpipe.
Compared to Mesa 25 it seems to round down instead of to nearest in
texture filtering, causing a really big difference in distance field and
blur output. As increasing thresholds globally to fit this would likely
cause rendering bugs to potentially go unnoticed but I still want to
test on the old llvmpipe on the CI, the affected comparisons are now an
XFAIL on old Mesa versions together with a more relaxed compare.