Ui: pass also masks of which animations started/stopped to advance().
Until now, the only ways to know whether a particular animation finished
playing during an advance() call, for example to perform a cleanup, was
to:

 - Either check factor for being exactly 1.0f. But that could give a
   false positive in case of repeated animations, where it could happen
   that the interpolation factor was calculated to be exactly 1.0f.
 - Or perform the cleanup in clean() instead. But that only works for
   animations that don't have KeepOncePlayed set.

Similarly, there was no way to check whether an animation just
now started playing, to perform some initial setup, and the only way was
to track that information manually by setting some "active" bit in the
animator implementations. Since this information is known internally,
let's just pass it to the interfaces.

This is a breaking change, affecting basically any custom animators
implemented so far. Sorry about that, but I have to do this in order to
be able to fix various blockers preventing implicit style animations,
node animators and generic "single shot" animators from working.