Ui: rename animation "played" to "start" or "started".
It's an extremely confusing term, not sure why I picked it in the first
place. Usually "played" is thought of as "something started and also
already finished playing by given time" while here it's used to mean "to
start playing at given time".

Furthermore, in case of create() I'm changing it from
create(played, duration) to create(start, duration) because the former
is just a complete nonsense that only starts making sense once you look
at examples of *what* is actually being passed as the first parameter.

This is a breaking change for the played() getter, which is now named
started(), but that one was likely only used in tests so I don't think
it breaks anything in practice. All other use cases are argument names,
docs and internal code, which won't cause breakages.