Ui: allow animations with duration being zero.
Originally I didn't want those because I wasn't sure what to pass as an
interpolation factor for them, or what to do in presence of a repeat.
But they're very useful for single-shot callbacks or generally
animations that just do something exactly once, and that cannot be
achieved otherwise, as using a 1-nanosecond duration can still result in
it being advanced twice if it fits exactly on the boundary of two
frames.

Zero-duration animations now enforce repeat count to be 1 for clarity
even though internally it doesn't really matter. Because people could
mistake those for "fire some callback regularly" if repeats were
allowed, which isn't the case.