Ui: doc++
Of course Doxygen just utterly failed about creating a link to that
function, and didn't tell me so I discovered it only months later. The
note about [0, 1] range was also incorrect, as the easing function is
free to do anything, even go out of bounds. The stray semicolon after an
ellipsis was also an error.