Ui: slightly adjust the NodeAnimator tests to not hit an edge case.
It was failing the Emscripten test because it interpolated from a NaN
value. I wanted to blame Emscripten for being shitty (because the local
version didn't fail), but ultimately it was my bug. And fixing that bug
uncovered another bug in the test, where it was testing for a thing that
shouldn't happen.

Thanks Emscripten, I guess?