Ui: make NodeAnimator aware of AnimationFlag::Reverse.
Not testing for ReverseEveryOther, as it only affects the factor value
and nothing else.

Originally I thought I'd have to add some addFlagsDuring() /
clearFlagsDuring() that'd allow for making animations that are actually
making sense when reversed, but only when writing the actual
documentation example I realized that it's enough to just reverse the
behavior of addFlagsBegin() etc. *enough*. Yay simplicity.