Ui: doc++, minor cleanup.
It makes more sense to have the DebugIntegration documented on the
AbstractLayer already, to not falsely hint that it's only possible on
DebugLayer sublcasses or some such.

Also, what the hell was the confusing wording for the debug layer
integration everywhere. Looks like every second word was written by a
different person and then it got copypasted several times.