Ui: test AbstractVisualLayerStyleAnimator::doClean() directly.
So far it was only tested as a consequence of an animation stopping and
being removed, but the behavior there will change to perform dynamic
style recycling on stop instead of on remove, meaning the
subsequent doClean() will be a no-op because there's nothing left to
recycle.

So test it directly to verify the code there continues to work, in
particular by removing data with animations attached and verifying that
the dynamic styles get recycled appropriately.