Ui: fetch {Base,Text}LayerStyleAnimator data at animation start.
Not at animation creation time, so existing animations can be reused
even after switching styles. So far this doesn't deal with appropriately
updating running animations to a style switch, but I can deal with that
later, for example by attaching to the style update stamp in the
animator and doing the same as for `started[i]` if it differs.

This change also means that it's no longer possible to query source and
target style properties for each animation, as those are only available
during the animation being played, and before or after they'd be
inevitably stale. Instead of wrapping the getters in some Optional, I
just removed them all and provided just a styles() getter.

The second return value from styles() is currently the same as
targetStyle(), but in the future they'll be likely independent, as
someone might want to for example animate a cursor of an active input
field by transitioning between two "edge" styles, but the style for when
the focus isn't on the field would be something completely different.