Ui: make AbstractLayerStyleAnimator aware of source style index also.
Originally I didn't put it there because the BaseLayer didn't need to
store such information. But now as fetching style data will be deferred
to animation start instead of animation creation in order to make the
animations better reusable, it has to store that index as well, and so
will any other animator, such as for the LineLayer. Which means it makes
sense to put this directly into the base class to reduce needless code
duplication. Similarly, in order to make the animators handle reversible
animations correctly, the base will need access to the source index as
well.

This is a breaking change in that the targetStyle() no longer exists and
there's instead a styles() getter, but as those were mainly used in
tests and not by any real-world code, I don't think they'll cause a
breakage in practice.