Ui: make style animators robust against external style switching.
They now remember the current style on animation start and perform a
switch to the dynamic style or to the target style only if the style is
as expected when the switch is about to happen. Should resolve most
cases of animations stomping over each other, especially in cases when
they're fired shortly after each other in response to input events.