Ui: improve treatment of (forever) scheduled animations.
Creating animations with start time set to Nanoseconds::max() was
documented as "creating a stoppped animation", however the implicit
result of doing that was that the animation was removed right after the
next advance. Which is most likely not wanted.

Furthermore, with 9149a3f0b8b9fab663d14635fe1263b08f4bf144, such
animations got advanced as if they've been already played in full,
setting both started and stopped bits, which is even worse, as
scheduling an animation 200 years in the future should definitely not
play it *now*. Similar case is now with animations that transition from
a scheduled to a stopped state.

First I attempted to create a special case for this, where animations
with the start time at max() would be treated as Scheduled instead of
Stopped, unfortunately that meant creating any such animation would
trigger NeedsAdvance from that point onwards even though the animation
is never going to actually start. And fixing that needed another special
case, making this even worse. Treating those as "Stopped but not
actually advancing for the final time or removing" was another attempt,
but that required another bunch of special casing in create() and
update(), and one more special case to not have them removed, and when I
attempted to document what is this doing, I realized the user would also
have hard time figuring out why there's a Stopped animation that's
treated in some weird special way. Finally, I realized the only way to
make this clear is to introduce AnimationState::Reserved, which gets
used only for animations that have the start time set to
Nanoseconds::max(), and which thus neither get affected by advance() nor
get removed as no longer used.

This commit then also contains corresponding changes to DebugLayer
printing and respective tests.