Ui: support for animating styles and style transitions.
Similarly to setStyleTransition(), setStyleAnimation() takes a set of
functions which can create animations for a particular style transition
(such as a fade out after the pointer leaves a widget) and/or a
persistent animation for the target style, such as a cursor animation.
Priority and correct order of created animations is enforced partially
by the robustness features added in previous commits, and partially by
assertions verifying various properties of created animations.