Ui: polish flat button and input corner radii for better transitions.
Currently if the animations faded to a zero-width outline, the inner
coner radius was interpolated towards 0, resulting in the corners being
cut off. Now not anymore.

For some reason, even though this should have resulted in no difference
whatsoever, for inputs with SubdividedQuads enabled this causes a slight
increase in the differences for the ARM64 build that uses Mesa 21.2.6.
Not natively on Mesa 25, and not on Mesa 20 used by other jobs, and
given it's with SubdividedQuads, I assume it's just some temporary
rounding difference that isn't worth further investigation.