Ui: rename DebugLayer::NodeHighlight and related APIs to NodeInspect.
Because it doesn't just highlight a node, it also inspects its details.
And I'm going to introduce a feature that just highlights, either
directly or based on some condition, so this has to be renamed. Some of
the internal names are kept as "highlight" as they'll be used for
general highlighting, not just inspection.

It's a breaking change, yes, sorry, but this functionality is still
relatively new so I don't think it's going to affect hundreds of
projects.