Ui: add DebugLayerFlag::NodeInspectSkipNoData.
I have a real-world use case where a rogue oversized but otherwise empty
node obscures a large portion of the application, making it impossible
to inspect what's underneath. Not yet sure how common this case might
be, but having an annoyance-removal feature for such cases might be
handy.

Eventually it might be useful to expand this to skip if it has no visual
or event data (but can have others), for now this is enough.