Ui: make layers, layouters and animators convertible to their handles.
This makes a lot of common use cases much nicer without having to
explicitly introduce overloads for all APIs that take a handle. Similar
thing is already done with widgets that are convertible to their
handles, so this feels like an obvious extension of that workflow.

The only place where I left the original explicit handle() call is in
various cases where handles are being compared in tests or with custom
handle creation and extraction.