Ui: simplify logic for layer/layouter/animator state propagation.
Put the max possible state to a variable and check against it to ensure
there aren't any copypaste errors.

Additionally, the early return was using == instead of a >=, causing
needless iteration of further layers, layouters and animators even
though the max state was already reached. Can't really test for
unnecessary calls to state() on those, so just fixing it.