Ui: extract thorough layouter update verification to a dedicated test.
There was no test for the actual layouter update in
AbstractUserInterfaceTest, it was only tested as part of state(), which
made it extremely verbose because the complex interaction between the
two layouters had to be repeatedly tested several times instead of just
once. The state() case was already over 4k lines long, so any
simplification is welcome.