[wip] Ui: introduce LayerFeature::Layout.
The original idea I had was to add an ability to specify min/max size
for each node, which the layouter could then pick up, because
specifying that directly on the layouter would tie such information to a
particular layouter implementation.

But then I realized having just a single piece of information per node
isn't sufficient. For example, a node containing a text might want to
get sized according to the text, so the TextLayer should somehow be able
to specify the constraint. And there might be more than one text, and it
should pick the maximum of all for example.

The final realization was that styles themselves may want to specify
minimal widget sizes for example, or various paddings and such. Which
resulted in what's in this commit -- each layer can now advertise that
it supplies layout properties, and in its doLayout() implementation
adjusts min, max sizes, aspect ratios, or paddings/margins of nodes it's
attached to. Those properties are then subsequently fed to the
layouters, which can (but don't have to) make use of these constraints.

All this information (52 bytes per node) is only fed to the layouter to
produce absolute node offsets and sizes, and isn't needed otherwise.
Thus it doesn't have to be a persistent state, but rather is calculated
on-the-fly only when layouts change. Which needs quite extensive changes
to the state tracking, but so is life.

Additionally, because querying the constraints might be expensive if
done on all data in the whole UI, doLayout() is supplied a mask of only
the data which are attached to visible nodes that have layouts assigned.
If a node has no layouts, there's nothing that could make use of such
constraints, so it's useless to calculate them in the first place.

A LayoutLayer, which allows widget developers to specify various styles
for paddings, sizes and such, is coming next.

TODO: there's that extra if that isn't well tested yet (some "update
  empty" test variant should cover this i hope?)
TODO: also finish the part where it doesn't call doLayout() at all if
  there are no layouters (right now it'd call it with all ezros ...
  which seems like a waste, say, if a pixel-perfect positioned project
  uses the TextLayer, it'd have to pay for the cost of doLayout() being
  called on all million texts with none of them used)