Ui: introduce LayoutLayer.
Makes it possible to supply style-driven layout properties for nodes
such as min/max size, aspect ratio, padding or margin. No high-level
docs yet, as actual usage of this layer depends on there being a support
for layout properties in the SnapLayouter first, so it can be used in
doc snippets and example screenshots.