Ui: take node padding/margin properties into account in SnapLayouter.
This replaces the layouter-global padding and margin, which was more of
a crutch than a real solution. I didn't see a reason to preserve those
further, so the setPadding() / setMargin() APIs are removed now, being a
breaking change.

In the previous commits the global padding/margin got turned into
LayoutLayer styles and thus there should be no visible change if using
the builtin widgets.