Ui: don't pass node parents to layouter update().
This got introduced in 1f76e25d021cd6d864b32fc8eae1d9dfe8602e2f to
(hopefully) support the use case in SnapLayouter. But it turns out that
having the layout hierarchy completely independent from node hierarchy
is not a good idea as it causes massive headaches with ownership and
basically anything else.

So instead I've been tigtening up the rules to make layouts match node
hierarchy, and with LayouterFeature::UniqueLayouts it's possible to
query node parents as well as a layout assigned to node parent directly
from layouter add() implementation, making it possible to do various
sanity checks directly there instead of deferring them to update() and
losing all stack trace information if anything goes wrong.

Finally, from an efficiency perspective, since node parents are
immutable and the mapping between nodes and layouts isn't 1:1, it's
significantly better to perform a parent layout query directly in add()
and then cache that information, instead of rebuilding the information
from scratch on every update(). That's mainly an issue when given
layouter affects just a small subset of the node tree, because with the
nodeParents it's several extra indirections:

  querying layout node assignment -> querying node parent -> querying
  layout assigned to given node

Merely supplying the nodeParents argument would be encouraging such
random memory access patterns, which is wrong. By not providing this
information at update() time it forces the implementations to do the
mapping differently and more efficient.