Ui: ignore node margin when snapping to a center inside.
This used to be okay back when margin on both sides had to be the same,
but with the ability to have uneven margins it now looks like a clear
mistake. Centering an element with uneven margins would not place it in
the center of containing edge, but slightly shifted, with no clear
reason why.